requirements.txt
pandas
numpy
scikit-learn
azure-identity
azure-ai-ml
pyarrow